+++ Common/Utilities/InvariantDictionary.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Pathoschild.Stardew.Common.Utilities
 {
-    /// <summary>An implementation of <see cref="Dictionary{TKey,TValue}"/> whose keys are guaranteed to use <see cref="StringComparer.InvariantCultureIgnoreCase"/>.</summary>
+    /// <summary>An implementation of <see cref="Dictionary{TKey,TValue}"/> whose keys are guaranteed to use <see cref="StringComparer.OrdinalIgnoreCase"/>.</summary>
     internal class InvariantDictionary<TValue> : Dictionary<string, TValue>
     {
         /*********
@@ -11,17 +11,17 @@ namespace Pathoschild.Stardew.Common.Utilities
         *********/
         /// <summary>Construct an instance.</summary>
         public InvariantDictionary()
-            : base(StringComparer.InvariantCultureIgnoreCase) { }
+            : base(StringComparer.OrdinalIgnoreCase) { }
 
         /// <summary>Construct an instance.</summary>
         /// <param name="values">The values to add.</param>
         public InvariantDictionary(IDictionary<string, TValue> values)
-            : base(values, StringComparer.InvariantCultureIgnoreCase) { }
+            : base(values, StringComparer.OrdinalIgnoreCase) { }
 
         /// <summary>Construct an instance.</summary>
         /// <param name="values">The values to add.</param>
         public InvariantDictionary(IEnumerable<KeyValuePair<string, TValue>> values)
-            : base(StringComparer.InvariantCultureIgnoreCase)
+            : base(StringComparer.OrdinalIgnoreCase)
         {
             foreach (var entry in values)
                 this.Add(entry.Key, entry.Value);