+++ LookupAnything/DrawHelper.cs
@@ -77,7 +77,7 @@ namespace Pathoschild.Stardew.LookupAnything
                     // split on newlines
                     string wordPart = word;
                     int newlineIndex;
-                    while ((newlineIndex = wordPart.IndexOf(Environment.NewLine, StringComparison.InvariantCulture)) >= 0)
+                    while ((newlineIndex = wordPart.IndexOf(Environment.NewLine, StringComparison.Ordinal)) >= 0)
                     {
                         if (newlineIndex == 0)
                         {